More warning fixes

+++ uae4all_gp2x_0.7.2a/src/filesys.cpp
@@ -61,7 +61,7 @@
 #define DUMPLOCK(u,x)
 #endif
 
-static long dos_errno(void)
+static int dos_errno(void)
 {
     int e = errno;
